migration/multifd: Add mapped-ram support to fd: URI
commit decdc76772c453ff1444612e910caa0d45cd8eac upstream.

If we receive a file descriptor that points to a regular file, there's
nothing stopping us from doing multifd migration with mapped-ram to
that file.

Enable the fd: URI to work with multifd + mapped-ram.

Note that the fds passed into multifd are duplicated because we want
to avoid cross-thread effects when doing cleanup (i.e. close(fd)). The
original fd doesn't need to be duplicated because monitor_get_fd()
transfers ownership to the caller.

Intel-SIG: commit decdc76772c4 migration/multifd: Add mapped-ram support to fd: URI

Signed-off-by: Fabiano Rosas <farosas@suse.de>
Reviewed-by: Peter Xu <peterx@redhat.com>
Link: https://lore.kernel.org/r/20240229153017.2221-23-farosas@suse.de
Signed-off-by: Peter Xu <peterx@redhat.com>

 Conflicts:
	migration/multifd.c
[jz: resolve context conflict]
Signed-off-by: Jason Zeng <jason.zeng@intel.com>